third_party/blink/web_tests/external/wpt/workers/constructors/SharedWorker/setting-port-members-expected.txt

This is a testharness.js-based test.
PASS postMessage
PASS start
PASS close
FAIL onmessage assert_equals: {handleEvent:() => {}} expected null but got object "[object Object]"
PASS addEventListener
PASS removeEventListener
PASS despatchEvent
Harness: the test ran to completion.
